requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9

torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
matplotlib==3.9.0
einops==0.7.0
timm==0.9.16

pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
